% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{AuditConfig}
\alias{AuditConfig}
\title{Cloud User Accounts API Objects 
Creates and manages users and groups for accessing Google Compute Engine virtual machines.}
\usage{
AuditConfig(exemptedMembers = NULL, service = NULL)
}
\arguments{
\item{exemptedMembers}{Specifies the identities that are exempted from 'data access' audit logging for the `service` specified above}

\item{service}{Specifies a service that will be enabled for 'data access' audit logging}
}
\value{
AuditConfig object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:00:35
filename: /Users/mark/dev/R/autoGoogleAPI/googleclouduseraccountsvmalpha.auto/R/clouduseraccounts_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AuditConfig Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Enables 'data access' audit logging for a service and specifies a list of members that are log-exempted.
}

