% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{AuthorizedKeysView}
\alias{AuthorizedKeysView}
\title{AuthorizedKeysView Object}
\usage{
AuthorizedKeysView(keys = NULL, sudoer = NULL)
}
\arguments{
\item{keys}{[Output Only] The list of authorized public keys in SSH format}

\item{sudoer}{[Output Only] Whether the user has the ability to elevate on the instance that requested the authorized keys}
}
\value{
AuthorizedKeysView object
}
\description{
AuthorizedKeysView Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of authorized public keys for a user account.
}
