% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{Condition}
\alias{Condition}
\title{Condition Object}
\usage{
Condition(iam = NULL, op = NULL, svc = NULL, sys = NULL, value = NULL,
  values = NULL)
}
\arguments{
\item{iam}{Trusted attributes supplied by the IAM system}

\item{op}{An operator to apply the subject with}

\item{svc}{Trusted attributes discharged by the service}

\item{sys}{Trusted attributes supplied by any service that owns resources and uses the IAM system for access control}

\item{value}{The object of the condition}

\item{values}{The objects of the condition}
}
\value{
Condition object
}
\description{
Condition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A condition to be met.
}
