% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{Group}
\alias{Group}
\title{Group Object}
\usage{
Group(creationTimestamp = NULL, description = NULL, id = NULL,
  members = NULL, name = NULL, selfLink = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional textual description of the resource; provided by the client when the resource is created}

\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{members}{[Output Only] A list of URLs to User resources who belong to the group}

\item{name}{Name of the resource; provided by the client when the resource is created}

\item{selfLink}{[Output Only] Server defined URL for the resource}
}
\value{
Group object
}
\description{
Group Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Group resource.
}
\seealso{
Other Group functions: \code{\link{groups.insert}}
}
