% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{LinuxGetAuthorizedKeysViewResponse}
\alias{LinuxGetAuthorizedKeysViewResponse}
\title{LinuxGetAuthorizedKeysViewResponse Object}
\usage{
LinuxGetAuthorizedKeysViewResponse(resource = NULL)
}
\arguments{
\item{resource}{[Output Only] A list of authorized public keys for a user}
}
\value{
LinuxGetAuthorizedKeysViewResponse object
}
\description{
LinuxGetAuthorizedKeysViewResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
