% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{TestPermissionsRequest}
\alias{TestPermissionsRequest}
\title{TestPermissionsRequest Object}
\usage{
TestPermissionsRequest(permissions = NULL)
}
\arguments{
\item{permissions}{The set of permissions to check for the 'resource'}
}
\value{
TestPermissionsRequest object
}
\description{
TestPermissionsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TestPermissionsRequest functions: \code{\link{groups.testIamPermissions}},
  \code{\link{users.testIamPermissions}}
}
