% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_functions.R
\name{users.get}
\alias{users.get}
\title{Returns the specified User resource.}
\usage{
users.get(project, user)
}
\arguments{
\item{project}{Project ID for this request}

\item{user}{Name of the user resource to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/cloud.useraccounts
\item https://www.googleapis.com/auth/cloud.useraccounts.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only, https://www.googleapis.com/auth/cloud.useraccounts, https://www.googleapis.com/auth/cloud.useraccounts.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/compute/docs/access/user-accounts/api/latest/}{Google Documentation}
}

